pyproject.toml
[project]
name = "goodcup-django"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django>=5.2.6",
    "dotenv>=0.9.9",
    "pillow>=11.3.0",
    "pydantic>=2.11.9",
    "pyjwt[crypto]>=2.10.1",
]


[tool.mypy]
igmore_missing_imports = true
disallow_incomplete_defs = true
disallow_untyped_calls = true
disallow_untyped_defs = true
exclude = [".venv"]
